Move defaultEventNames into Schema
This allows applications to extend the mapping with support for custom
elements.

Closes #660
