+++ src/core/schema.ts
@@ -5,6 +5,7 @@ export interface Schema {
   targetAttributeForScope(identifier: string): string
   outletAttributeForScope(identifier: string, outlet: string): string
   keyMappings: { [key: string]: string }
+  defaultEventNames: { [tagName: string]: (element: Element) => string }
 }
 
 export const defaultSchema: Schema = {
@@ -29,6 +30,22 @@ export const defaultSchema: Schema = {
     // [0-9]
     ...objectFromEntries("0123456789".split("").map((n) => [n, n])),
   },
+  defaultEventNames: {
+    a: () => "click",
+    button: () => "click",
+    form: () => "submit",
+    details: () => "toggle",
+    input: (e) => (e.getAttribute("type") == "submit" ? "click" : "input"),
+    select: () => "change",
+    textarea: () => "input",
+  }
+}
+
+export function getDefaultEventNameForElement(element: Element, schema = defaultSchema): string | undefined {
+  const tagName = element.tagName.toLowerCase()
+  if (tagName in schema.defaultEventNames) {
+    return schema.defaultEventNames[tagName](element)
+  }
 }
 
 function objectFromEntries(array: [string, any][]): object {
